same url for admin and public #done
