+++ app/Http/Controllers/App/AppController.php
@@ -17,7 +17,7 @@ class AppController extends Controller
      */
     public function index()
     {
-        return $this->view->make('app.app', [
+        return view('app.app', [
             'pusherId'           => env('PUSHER_APP_KEY'),
             'public_bridge_path' => url('/') . "/bridge/",
             'is_public'          => false,
@@ -35,6 +35,13 @@ class AppController extends Controller
     {
         $bridge = Bridge::with('sections', 'icons', 'icons.converted', 'images', 'images.converted', 'fonts', 'fonts.variant', 'fonts.variant.fontFamily', 'colors')->where('slug', $slug)->first();
 
+        $loggedUser = auth()->user();
+        if ($loggedUser) {
+            $isPublic = $loggedUser->id === $bridge->user_id;
+        } else {
+            $isPublic = false;
+        }
+
         if ($bridge->count() === 0) {
             throw new ModelNotFoundException;
         }
@@ -42,27 +49,9 @@ class AppController extends Controller
         return view('app.project', [
             'pusherId'           => env('PUSHER_APP_KEY'),
             'public_bridge_path' => url('/') . "/",
-            'is_public'          => auth()->check() ? false : true,
+            'is_public'          => $isPublic ? false : true,
             'bridge'             => json_encode($bridge),
             'section_types'      => json_encode(SectionType::all()),
         ]);
     }
-
-    public function publicIdentities($slug)
-    {
-        $bridge = Bridge::with('sections', 'icons', 'icons.converted', 'images', 'images.converted', 'fonts', 'fonts.variant', 'fonts.variant.fontFamily', 'colors')->where('slug', $slug)->get();
-
-        if ($bridge->count() === 0) {
-            throw new ModelNotFoundException;
-        }
-
-        return $this->view->make('app.public', [
-            'pusherId'           => env('PUSHER_APP_KEY'),
-            'public_bridge_path' => url('/') . "/identities/",
-            'is_public'          => true,
-            'bridge'             => json_encode($bridge->first()),
-            'section_types'      => json_encode(SectionType::all()),
-        ]);
-    }
-
 }
