+++ public/mix-manifest.json
@@ -1,6 +1,5 @@
 {
-    "/js/react-panel.js": "/js/react-panel.js?id=feeeef95765636d618d3",
-    "/css/app.css": "/css/app.css?id=2e476dafcd9973ce9ce3",
-    "/js/react-public.js": "/js/react-public.js?id=d7b9ab6a1103d7c31d95",
-    "/js/react-project.js": "/js/react-project.js?id=9297728eb37f7d39ec6b"
+    "/js/react-project.js": "/js/react-project.js?id=4f11b0b6c2b295ca255d",
+    "/js/react-panel.js": "/js/react-panel.js?id=afd6aefd6c7e084aef28",
+    "/css/app.css": "/css/app.css?id=2e476dafcd9973ce9ce3"
 }
