+++ resources/assets/js/components/bridge/BridgesEmptyState.js
