+++ resources/assets/js/routes.js
@@ -1,11 +1,7 @@
 import React from 'react';
-import {HashRouter, Route, Switch, Redirect} from 'react-router-dom';
+import {HashRouter, Route, Redirect} from 'react-router-dom';
 import BridgesList from './pages/BridgesPage';
 import NewBridge from './pages/NewBridge';
-import Bridge from './pages/Bridge';
-import AddColor from './pages/AddColor';
-import AddFont from './pages/AddFont';
-import Viewer from './pages/viewer/Viewer';
 import DeleteBridge from './pages/DeleteBridge';
 import Helmet from 'react-helmet';
 
@@ -21,10 +17,6 @@ export default (
                 <Route path="/projects" strict={true} component={BridgesList}/>
                 <Route path="/projects/:id/delete-confirmation" component={DeleteBridge}/>
                 <Route path="/projects/new" strict={true} component={NewBridge}/>
-                <Route path="/project/:id" component={Bridge}/>
-                <Route path="/project/:id/add-color" component={AddColor}/>
-                <Route path="/project/:id/add-font" component={AddFont}/>
-                <Route path="/project/:id/view/:objectType/element/:elementId" component={Viewer}/>
             </div>
         </HashRouter>
     </div>
