--- resources/views/app/public.blade.php
@@ -1,9 +1,0 @@
-@extends('layouts.public')
-
-@section('content')
-
-    <div class="bridges big-with-margin-top container">
-        <div id="app"></div>
-    </div>
-
-@endsection
