+++ webpack.mix.js
@@ -13,7 +13,6 @@ const {mix} = require('laravel-mix');
 
 mix.sass('resources/assets/sass/app.scss', 'public/css').version();
 mix.react('resources/assets/js/index.js', 'public/js/react-panel.js').version();
-mix.react('resources/assets/js/public-bridge.js', 'public/js/react-public.js').version();
 mix.react('resources/assets/js/modules/project/index.js', 'public/js/react-project.js').version();
 
 mix.disableNotifications();
